apps/api/requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.8.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.8
celery==5.4.0
httpx==0.27.0
reportlab==4.2.2
ruff==0.5.7
black==24.8.0
pytest==8.3.2
loguru==0.7.2
